BDOG-901: Update to use -r instead of -f for runFromBinary
The -f flag looks for snapshot builds, whereas -r looks for release
builds. Since everything that is built by Jenkins is now a release (and
has been for 2 years?) it makes sense to default to this.

Furthermore, the -f will be deprecated in the near future for newer
versions of service-manager.